fix: Item 수정 시 이미지 String으로 받기
