+++ src/main/java/com/rudkids/rudkids/domain/item/ItemFactory.java
@@ -4,4 +4,6 @@ import com.rudkids.rudkids.domain.item.domain.Item;
 
 public interface ItemFactory {
     Item create(ItemCommand.CreateItemRequest command);
+
+    void update(Item item, ItemCommand.UpdateRequest command);
 }
