+++ src/main/java/com/rudkids/rudkids/infrastructure/item/ItemFactoryImpl.java
@@ -5,6 +5,8 @@ import com.rudkids.rudkids.domain.item.ItemFactory;
 import com.rudkids.rudkids.domain.item.domain.*;
 import org.springframework.stereotype.Component;
 
+import java.util.stream.Collectors;
+
 @Component
 public class ItemFactoryImpl implements ItemFactory {
 
@@ -21,5 +23,18 @@ public class ItemFactoryImpl implements ItemFactory {
             .map(img -> ItemImage.create(initItem, img.path(), img.url()))
             .forEach(initItem::addImage);
         return initItem;
+    }
+
+    @Override
+    public void update(Item item, ItemCommand.UpdateRequest command) {
+        var name = Name.create(command.name());
+        var itemBio = ItemBio.create(command.itemBio());
+        var price = Price.create(command.price());
+        var quantity = Quantity.create(command.quantity());
+        var limitType = command.limitType();
+        command.images().stream()
+            .map(img -> ItemImage.create(item, img.path(), img.url()))
+            .forEach(item::addImage);
+        item.update(name, itemBio, price, quantity, limitType);
     }
 }
